Open staff schedules on dedicated page
